Undoing some changes in dequantize.

+++ src/quantization/common.jl
@@ -47,8 +47,7 @@ function dequantize!(y::AbstractVector{Float32}, x::AbstractVector{Float32})
 end
 
 function dequantize(x::AbstractVector{<:Union{block_q4_K, block_q5_K, block_q6_K, block_q8_K}})
-    y = zeros(Float16, length(x)*QK_K) # TODO use this!
-    # y = zeros(Float32, length(x)*QK_K)
+    y = zeros(Float32, length(x)*QK_K)
     dequantize!(y, x)
     return y
 end
